move abtestexample to animations
