+++ src/components/Activities/Activity3/Animations/IntroABTestExplanations.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import ABCompare from '../../../assets/activity3/AB-compare.svg';
-import ABGraph from '../../../assets/activity3/AB-graph.svg';
+import ABCompare from '../../../../assets/activity3/AB-compare.svg';
+import ABGraph from '../../../../assets/activity3/AB-graph.svg';
 
 export function ABTestSlide1(): JSX.Element {
   return <img src={ABCompare} alt='version a versus version b' style={{ maxHeight: '50vh' }} />;
